fix: rename the discount_lodash modules in case of special plugins looking for 'lodash' in the module names
