+++ packages/core/react-dnd/src/decorators/DragSource.ts
@@ -2,7 +2,7 @@ import * as React from 'react'
 import invariant from 'invariant'
 import { SourceType, DragDropManager } from 'dnd-core'
 import { DndOptions } from '../interfaces'
-import { isPlainObject } from '../utils/discount_lodash'
+import { isPlainObject } from '../utils/js_utils'
 import {
 	DndComponentEnhancer,
 	DragSourceSpec,
