+++ packages/core/react-dnd/src/decorators/DropTarget.ts
@@ -2,7 +2,7 @@ import * as React from 'react'
 import { TargetType, DragDropManager } from 'dnd-core'
 import invariant from 'invariant'
 import { DndOptions } from '../interfaces'
-import { isPlainObject } from '../utils/discount_lodash'
+import { isPlainObject } from '../utils/js_utils'
 import { registerTarget } from '../common/registration'
 import { isValidType } from '../utils/isValidType'
 import { TargetConnector } from '../common/TargetConnector'
